requirements.txt
django>=3.1,<4.0
djangorestframework>=3.12,<4.0
